% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cluster.R
\name{k_cluster}
\alias{k_cluster}
\title{A cluster algorithm in K_means}
\usage{
k_cluster(x, y, n)
}
\arguments{
\item{x}{first dimension of the pointset}

\item{y}{second dimension of the pointset}

\item{n}{number of clusters}
}
\value{
the belonged cluster of each point
}
\description{
\code{PESQ_mos} and \code{MOS_lqo} are vital indexes in the period of Speech Separation with a funtcional relationship
}
\examples{
\dontrun{
k_cluster(cars$dist, cars$speed, 3)
}
}
