% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{SS}
\alias{SS}
\title{Using R to give a transform between MOS and PESQ}
\usage{
SS(type, value)
}
\arguments{
\item{type}{determine the received type of index,"MOS_lqo" or "PESQ_mos"(don't forget double quotes)}

\item{value}{the value of input index, should between (1,4.5)}
}
\value{
a transformed value of index (\code{PESQ_mos} or \code{MOS_lqo})
}
\description{
\code{PESQ_mos} and \code{MOS_lqo} are vital indexes in the period of Speech Separation with a funtcional relationship
}
\examples{
\dontrun{
SS("MOS_lqo",seq(1, 4.5, by=0.1))
}
}
